custom button views, layout

+++ app/src/main/java/com/qwerty/tetriz/Grid.java
@@ -66,10 +66,8 @@ public class Grid
         int width = canvas.getWidth();
         int height = canvas.getHeight();
 
-        spacing = (int) ((Math.min(width, height) / columns) * 1);
+        spacing = (Math.min(width, height) / columns) * 1;
 
-//        leftMargin = (width - spacing * (columns)) / 2;
-//        topMargin = (height - spacing * (rows)) / 2;
         //leftMargin = (int)(width*0.2);
         topMargin = (int)(height*0.1);
         drawVerticalLines(canvas);
